Solutions_ya/tp_4/requirements.txt
# Streamlit
streamlit

# API
fastapi
pydantic
uvicorn
pydantic
requests
langchain-google-genai
python-dotenv==1.0.1
langchain-google-cloud-sql-pg
langchain-google-vertexai
langchain_community
google-cloud-storage
unstructured
langchain
langchain_unstructured